Advisor: Check plugin signature
